move related files to ProGuide folder
