+++ src/components/views/create/proGuide/ProjectTips/BannerImageTip.tsx
