+++ src/components/views/create/proGuide/ProjectTips/common.styles.tsx
