+++ src/components/views/create/proGuide/common.sc.ts
