+++ src/components/views/create/proGuide/score/ScoreBox.tsx
